+++ src/main/java/com/culinary/userservice/recipe/controller/DietTypeController.java
@@ -22,19 +22,19 @@ public class DietTypeController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<DietTypeDTO> updateDietType(@PathVariable int id, @RequestBody DietTypeDTO dietTypeDTO) {
+    public ResponseEntity<DietTypeDTO> updateDietType(@PathVariable long id, @RequestBody DietTypeDTO dietTypeDTO) {
         DietTypeDTO updatedDietType = dietTypeService.updateDietType(id, dietTypeDTO);
         return ResponseEntity.ok(updatedDietType);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteDietType(@PathVariable int id) {
+    public ResponseEntity<Void> deleteDietType(@PathVariable long id) {
         dietTypeService.deleteDietType(id);
         return ResponseEntity.noContent().build();
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<DietTypeDTO> getDietTypeById(@PathVariable int id) {
+    public ResponseEntity<DietTypeDTO> getDietTypeById(@PathVariable long id) {
         DietTypeDTO dietTypeDTO = dietTypeService.getDietTypeById(id);
         return ResponseEntity.ok(dietTypeDTO);
     }
