+++ src/main/java/com/culinary/userservice/recipe/controller/ReviewController.java
@@ -23,19 +23,19 @@ public class ReviewController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ReviewDTO> updateReview(@PathVariable int id, @RequestBody PutReviewDTO updateReviewDTO) {
+    public ResponseEntity<ReviewDTO> updateReview(@PathVariable long id, @RequestBody PutReviewDTO updateReviewDTO) {
         ReviewDTO updatedReview = reviewService.updateReview(id, updateReviewDTO);
         return ResponseEntity.ok(updatedReview);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReview(@PathVariable int id) {
+    public ResponseEntity<Void> deleteReview(@PathVariable long id) {
         reviewService.deleteReview(id);
         return ResponseEntity.noContent().build();
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ReviewDTO> getReviewById(@PathVariable int id) {
+    public ResponseEntity<ReviewDTO> getReviewById(@PathVariable long id) {
         ReviewDTO reviewDTO = reviewService.getReviewById(id);
         return ResponseEntity.ok(reviewDTO);
     }
